# backend/TeamEventPlanner.Api/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "TeamEventPlanner.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Events": {
      "post": {
        "tags": [
          "Events"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateEventDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Events"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reports/data": {
      "get": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "end",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "clientLocalTime",
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reports/event-report": {
      "get": {
        "tags": [
          "Reports"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reports/Bold": {
      "get": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "end",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "clientLocalTime",
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "CreateAttendeeDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "CreateEventDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "startTime": {
          "format": "date-time",
          "type": "string"
        },
        "endTime": {
          "format": "date-time",
          "type": "string"
        },
        "venue": {
          "type": "string"
        },
        "attendees": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CreateAttendeeDto"
          }
        }
      },
      "additionalProperties": false
    }
  }
}